close connection on readtimeout
Pings will be sent at the interval defined by activityTimeout and a pong must be received within the pongTimeout. So websocket readDeadline will be set to activityTimeout + pongTimeout. (pongTimeout  hardcoded to 30 secs as recommended by Pusher API). If the websocket read returns an error, the connection is closed. Any registered listeners will be notified of the closed connection. The user will be responsible for the logic to reconnect the client by calling Connect() again.