add chan bool stopper to stop heartbeat goroutine on disconnect
This prevents a crash that happens when the websocket connection is disconnected. Heartbeat uses the mutex controlled check to see if the websocket isconnected, however, the `select` blocks until the activityTimer fires which could be seconds later, and the `c.ws` websocket object could already be nil’ed out causing a nil pointer dereference crash. This adds a chan bool stopper to quit the go routine on disconnect.