reduce scope of mutex lock to prevent deadlock
This reduces the scope of the mutex lock to just protect the mutex sensitive vars instead of the entire Unsubscribe() func. This was causing a deadlock because the `ch.Unsubscribe` calls SendEvent() on the client, which now checks to see if the websocket is connected which is grabs a mutex lock.